Numeric.Statistics.PCA

Description

Principal Components Analysis

Documentation

pca :: Array Int (Vector Double) -> Double -> Matrix Double

find the principal components of multidimensional data greater than the threshhold

pcaN :: Array Int (Vector Double) -> Int -> Matrix Double

find N greatest principal components of multidimensional data according to size of the eigenvalue

pcaTransform

Arguments

:: Array Int (Vector Double)
the data

-> Matrix Double
the principal components

-> Array Int (Vector Double)
the transformed data

perform a PCA transform of the original data (remove mean) | Final = M^T Data^T

pcaReduce

Arguments

:: Array Int (Vector Double)
the data

-> Double
eigenvalue threshold

-> Array Int (Vector Double)
the reduced data

perform a dimension-reducing PCA modification, using an eigenvalue threshhold

pcaReduceN

Arguments

:: Array Int (Vector Double)
the data

-> Int
N, the number of components

-> Array Int (Vector Double)
the reduced data, with n principal components

perform a dimension-reducing PCA modification, using N components