Numeric.Statistics.Surrogate

Description

Methods for tests using surrogate data

Documentation

surrogate

Arguments

:: Int
random seed

-> Int
number of repetitions

-> (Array Int (Vector Double) -> a)
the evaluation function

-> Array Int (Vector Double)
the data

-> Array Int a
the results, with the evaluated real data in position 1 and the rest of the array containing the evaluated surrogate data

perform an analysis using surrogate data